--- kubernetes-manifests/config.yaml
@@ -1,42 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: environment-config
-data:
-  LOCAL_ROUTING_NUM: "883745000"
-  PUB_KEY_PATH: "/root/.ssh/publickey"
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: service-api-config
-data:
-  TRANSACTIONS_API_ADDR: "ledgerwriter:8080"
-  BALANCES_API_ADDR: "balancereader:8080"
-  HISTORY_API_ADDR: "transactionhistory:8080"
-  CONTACTS_API_ADDR: "contacts:8080"
-  USERSERVICE_API_ADDR: "userservice:8080"
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: demo-data-config
-data:
-  USE_DEMO_DATA: "True"
-  DEMO_LOGIN_USERNAME: "testuser"
-  # All demo user accounts are hardcoded to use the login password 'password'
-  DEMO_LOGIN_PASSWORD: "password"
