--- src/ledgermonolith/kubernetes-manifests/accounts-db.yaml
@@ -1,90 +1,0 @@
-# Copyright 2019 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-kind: StatefulSet
-apiVersion: apps/v1
-metadata:
-  name: accounts-db
-  labels:
-    app: accounts-db
-    tier: db
-spec:
-  serviceName: "accounts-db"
-  replicas: 1
-  selector:
-    matchLabels:
-      app: accounts-db
-      tier: db
-  template:
-    metadata:
-      labels:
-        app: accounts-db
-        tier: db
-    spec:
-      containers:
-      - name: accounts-db
-        image: gcr.io/bank-of-anthos/accounts-db:latest
-        envFrom:
-          - configMapRef:
-              name: environment-config
-          - configMapRef:
-              name: accounts-db-config
-          - configMapRef:
-              name: demo-data-config
-        ports:
-          - containerPort: 5432
-            name: postgredb
-        resources:
-          requests:
-            cpu: 100m
-            memory: 128Mi
-          limits:
-            cpu: 500m
-            memory: 1Gi
-        volumeMounts:
-        - name: postgresdb
-          mountPath: /var/lib/postgresql/data
-          subPath: postgres
-      volumes:
-      - name: postgresdb
-        emptyDir: {}
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: accounts-db
-  labels:
-    app: accounts-db
-    tier: db
-spec:
-  ports:
-    - port: 5432
-      name: tcp
-      targetPort: 5432
-      protocol: TCP
-  selector:
-    app: accounts-db
-    tier: db
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: accounts-db-config
-  labels:
-    app: accounts-db
-data:
-  POSTGRES_DB: accounts-db
-  POSTGRES_USER: accounts-admin
-  POSTGRES_PASSWORD: accounts-pwd
-  ACCOUNTS_DB_URI: postgresql://accounts-admin:accounts-pwd@accounts-db:5432/accounts-db
