+++ src/ledgermonolith/scripts/deploy-monolith.sh
@@ -70,21 +70,3 @@ gcloud compute instances create ledgermonolith-service \
     --metadata-from-file startup-script=${CWD}/../init/startup-script.sh \
     --tags monolith \
     --quiet
-
-
-# Allow HTTP access to the VM via firewall rule if it doesn't already exist
-echo "Creating firewall rule..."
-gcloud compute firewall-rules describe allow-http-monolith \
-    --project $PROJECT_ID \
-    --quiet >/dev/null 2>&1 
-if [ $? -ne 0 ]; then
-  gcloud compute firewall-rules create allow-http-monolith \
-      --project $PROJECT_ID \
-      --network default \
-      --allow tcp:8080 \
-      --source-tags bank-of-anthos \
-      --target-tags monolith \
-      --description "Allow port 8080 from bank-of-anthos to monolith VMs" \
-      --quiet
-fi
-
