--- src/ledgermonolith/skaffold.yaml
@@ -1,39 +1,0 @@
-# Copyright 2019 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: skaffold/v2alpha4
-kind: Config
-build:
-  artifacts:
-  - image: gcr.io/bank-of-anthos/frontend
-    context: ../frontend
-  - image: gcr.io/bank-of-anthos/contacts
-    context: ../contacts
-  - image: gcr.io/bank-of-anthos/userservice
-    context: ../userservice
-  - image: gcr.io/bank-of-anthos/loadgenerator
-    context: ../loadgenerator
-  - image: gcr.io/bank-of-anthos/accounts-db
-    context: ../accounts-db
-  tagPolicy:
-    gitCommit: {}
-  local:
-    concurrency: 0
-deploy:
-  kubectl:
-    manifests:
-    - ./kubernetes-manifests/**.yaml
-    - ../../extras/jwt/jwt-secret.yaml
-profiles:
-  - name: local-code
